pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.39.0",
]
